# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_create_release_api_releases_post": {
        "properties": {
          "bundle": {
            "contentMediaType": "application/octet-stream",
            "title": "Bundle",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "bundle"
        ],
        "title": "Body_create_release_api_releases_post",
        "type": "object"
      },
...
      "CodeRef": {
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "end_line": {
            "title": "End Line",
            "type": "integer"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "quote": {
            "title": "Quote",
            "type": "string"
          },
          "start_line": {
            "title": "Start Line",
            "type": "integer"
          },
          "type": {
            "const": "code",
            "default": "code",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "artifact_id",
          "path",
          "start_line",
          "end_line",
          "quote"
        ],
        "title": "CodeRef",
        "type": "object"
      },
      "Confidence": {
        "properties": {
          "applicability": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Applicability",
            "type": "number"
          },
          "evidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Evidence",
            "type": "number"
          },
          "finding": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Finding",
            "type": "number"
          }
        },
        "required": [
          "applicability",
          "evidence",
          "finding"
        ],
        "title": "Confidence",
        "type": "object"
      },
...
      "FindingDetail": {
        "properties": {
          "finding": {
            "$ref": "#/components/schemas/FindingView"
          },
          "provisions": {
            "items": {
              "$ref": "#/components/schemas/LegalProvision"
            },
            "title": "Provisions",
            "type": "array"
          },
          "requirement": {
            "$ref": "#/components/schemas/Requirement"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/Review"
            },
            "title": "Reviews",
            "type": "array"
          }
        },
        "required": [
          "finding",
          "requirement"
        ],
        "title": "FindingDetail",
        "type": "object"
      },
      "FindingView": {
        "description": "A finding as served by the API: AI value plus the applicable review, if any.",
        "properties": {
          "applicable_review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Review"
              },
              {
                "type": "null"
              }
            ]
          },
          "assessment_id": {
            "title": "Assessment Id",
            "type": "string"
          },
          "attempts": {
            "default": 1,
            "title": "Attempts",
            "type": "integer"
          },
          "carried_from_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carried From Version"
          },
          "citations": {
            "items": {
              "type": "string"
            },
            "title": "Citations",
            "type": "array"
          },
          "conclusion": {
            "enum": [
              "satisfied",
              "potential_violation",
              "insufficient_evidence",
              "not_applicable",
              "uncertain"
            ],
            "title": "Conclusion",
            "type": "string"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "effective_conclusion": {
            "enum": [
              "satisfied",
              "potential_violation",
              "insufficient_evidence",
              "not_applicable",
              "uncertain"
            ],
            "title": "Effective Conclusion",
            "type": "string"
          },
          "evidence": {
            "items": {
              "discriminator": {
                "mapping": {
                  "code": "#/components/schemas/CodeRef",
                  "document_span": "#/components/schemas/DocumentSpanRef",
                  "missing": "#/components/schemas/MissingRef"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/DocumentSpanRef"
                },
                {
                  "$ref": "#/components/schemas/CodeRef"
                },
                {
                  "$ref": "#/components/schemas/MissingRef"
                }
              ]
            },
            "title": "Evidence",
            "type": "array"
          },
          "evidence_fingerprint": {
            "title": "Evidence Fingerprint",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reasoning_summary": {
            "title": "Reasoning Summary",
            "type": "string"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "string"
          },
          "severity": {
            "enum": [
              "blocker",
              "high",
              "medium",
              "low"
            ],
            "title": "Severity",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "validation_notes": {
            "items": {
              "type": "string"
            },
            "title": "Validation Notes",
            "type": "array"
          }
        },
        "required": [
          "id",
          "assessment_id",
          "requirement_id",
          "conclusion",
          "severity",
          "title",
          "reasoning_summary",
          "confidence",
          "evidence_fingerprint",
          "effective_conclusion"
        ],
        "title": "FindingView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImportRequest": {
        "properties": {
          "result": {
            "$ref": "#/components/schemas/CiResult"
          }
        },
        "required": [
          "result"
        ],
        "title": "ImportRequest",
        "type": "object"
      },
      "LegalProvision": {
        "properties": {
          "act_title": {
            "title": "Act Title",
            "type": "string"
          },
          "article": {
            "title": "Article",
            "type": "string"
          },
          "celex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Celex"
          },
          "embedding": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "issuer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer"
          },
          "jurisdiction": {
            "title": "Jurisdiction",
            "type": "string"
          },
          "kind": {
            "default": "law",
            "enum": [
              "law",
              "guidance"
            ],
            "title": "Kind",
            "type": "string"
          },
          "legi_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legi Id"
          },
          "paragraph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraph"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "source": {
            "enum": [
              "cellar",
              "legifrance",
              "manual"
            ],
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "jurisdiction",
          "act_title",
          "article",
          "text",
          "source_url",
          "retrieved_at"
        ],
        "title": "LegalProvision",
        "type": "object"
      },
      "MissingRef": {
        "properties": {
          "artifact_kind": {
            "enum": [
              "business_plan",
              "product_spec",
              "privacy_policy",
              "terms",
              "regulatory_registration",
              "code_repo",
              "other"
            ],
            "title": "Artifact Kind",
            "type": "string"
          },
          "type": {
            "const": "missing",
            "default": "missing",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "artifact_kind"
        ],
        "title": "MissingRef",
        "type": "object"
      },
...
      "ReleaseOut": {
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/Artifact"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "latest_assessment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Assessment"
              },
              {
                "type": "null"
              }
            ]
          },
          "release": {
            "$ref": "#/components/schemas/Release"
          }
        },
        "required": [
          "release"
        ],
        "title": "ReleaseOut",
        "type": "object"
      },
...
      "Review": {
        "properties": {
          "comment": {
            "default": "",
            "title": "Comment",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "enum": [
              "confirm",
              "override",
              "not_applicable",
              "need_evidence"
            ],
            "title": "Decision",
            "type": "string"
          },
          "evidence_fingerprint": {
            "title": "Evidence Fingerprint",
            "type": "string"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "override_conclusion": {
            "anyOf": [
              {
                "enum": [
                  "satisfied",
                  "potential_violation",
                  "insufficient_evidence",
                  "not_applicable",
                  "uncertain"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Conclusion"
          },
          "product_id": {
            "title": "Product Id",
            "type": "string"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "string"
          },
          "reviewer_name": {
            "title": "Reviewer Name",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          }
        },
        "required": [
          "id",
          "finding_id",
          "requirement_id",
          "product_id",
          "reviewer_name",
          "decision",
          "evidence_fingerprint",
          "created_at"
        ],
        "title": "Review",
        "type": "object"
      },
      "ReviewCreate": {
        "properties": {
          "comment": {
            "default": "",
            "title": "Comment",
            "type": "string"
          },
          "decision": {
            "enum": [
              "confirm",
              "override",
              "not_applicable",
              "need_evidence"
            ],
            "title": "Decision",
            "type": "string"
          },
          "override_conclusion": {
            "anyOf": [
              {
                "enum": [
                  "satisfied",
                  "potential_violation",
                  "insufficient_evidence",
                  "not_applicable",
                  "uncertain"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Conclusion"
          },
          "reviewer_name": {
            "title": "Reviewer Name",
            "type": "string"
          }
        },
        "required": [
          "reviewer_name",
          "decision"
        ],
        "title": "ReviewCreate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CCOmmit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/assessments/{assessment_id}/findings": {
      "get": {
        "operationId": "list_findings_api_assessments__assessment_id__findings_get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FindingView"
                  },
                  "title": "Response List Findings Api Assessments  Assessment Id  Findings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Findings"
      }
    },
...
    "/api/findings/{finding_id}": {
      "get": {
        "operationId": "get_finding_api_findings__finding_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Finding"
      }
    },
    "/api/findings/{finding_id}/reviews": {
      "post": {
        "operationId": "create_review_api_findings__finding_id__reviews_post",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Review"
      }
    },
...
    "/api/provisions/{provision_id}": {
      "get": {
        "operationId": "get_provision_api_provisions__provision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "provision_id",
            "required": true,
            "schema": {
              "title": "Provision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalProvision"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Provision"
      }
    },
    "/api/provisions/{provision_id}/related": {
      "get": {
        "operationId": "related_provisions_api_provisions__provision_id__related_get",
        "parameters": [
          {
            "in": "path",
            "name": "provision_id",
            "required": true,
            "schema": {
              "title": "Provision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LegalProvision"
                  },
                  "title": "Response Related Provisions Api Provisions  Provision Id  Related Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Related Provisions"
      }
    },
    "/api/releases": {
      "get": {
        "operationId": "list_releases_api_releases_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReleaseOut"
                  },
                  "title": "Response List Releases Api Releases Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Releases"
      },
      "post": {
        "operationId": "create_release_api_releases_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_release_api_releases_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Release"
      }
    },
    "/api/releases/import": {
      "post": {
        "operationId": "import_release_api_releases_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Release"
      }
    },
    "/api/releases/{release_id}": {
      "get": {
        "operationId": "get_release_api_releases__release_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "release_id",
            "required": true,
            "schema": {
              "title": "Release Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Release"
      }
    },
...
    "/api/reviews/{review_id}/revoke": {
      "post": {
        "operationId": "revoke_review_api_reviews__review_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Review"
      }
    },
...
  }
}